requirements.txt
discord.py
pony
PyMySQL
